requirements.txt
Django==4.2
django-ninja==0.21.0
isort==5.13.2
pydantic==1.10.13
